requirements.txt
GitPython
pandas
rich
